test: improves test organization
